LPD-35325 move Multishipping component to commerce-frontend-js
