+++ modules/apps/commerce/commerce-checkout-web/src/main/resources/META-INF/resources/checkout_step/view_order_summary_multishipping.jsp
@@ -14,7 +14,7 @@ CommerceOrder commerceOrder = orderSummaryCheckoutStepDisplayContext.getCommerce
 %>
 
 <react:component
-	module="{Multishipping} from commerce-order-content-web"
+	module="{Multishipping} from commerce-frontend-js"
 	props='<%=
 		HashMapBuilder.<String, Object>put(
 			"accountId", commerceOrder.getCommerceAccountId()
