+++ modules/apps/commerce/commerce-frontend-js/package.json
@@ -6,12 +6,15 @@
 		"@clayui/core": "3.121.0",
 		"@clayui/data-provider": "3.119.0",
 		"@clayui/drop-down": "3.121.0",
+		"@clayui/empty-state": "3.116.0",
 		"@clayui/form": "3.119.0",
 		"@clayui/icon": "3.111.0",
 		"@clayui/label": "3.111.0",
 		"@clayui/layout": "3.111.0",
+		"@clayui/link": "3.111.0",
 		"@clayui/list": "3.121.0",
 		"@clayui/loading-indicator": "3.111.0",
+		"@clayui/management-toolbar": "3.111.0",
 		"@clayui/modal": "3.119.0",
 		"@clayui/multi-select": "3.121.0",
 		"@clayui/pagination-bar": "3.121.0",
@@ -26,7 +29,9 @@
 		"@liferay/frontend-js-react-web": "*",
 		"@liferay/frontend-js-state-web": "*",
 		"classnames": "2.3.1",
+		"fetch-mock": "5.13.1",
 		"frontend-js-web": "*",
+		"lodash": "4.17.21",
 		"prop-types": "15.7.2",
 		"react": "16.12.0",
 		"react-dom": "16.12.0"
@@ -45,6 +50,9 @@
 			"<rootDir>/src/**/utilities/**/*.js",
 			"!<rootDir>/src/**/constants.js",
 			"!<rootDir>/src/**/entry.js"
+		],
+		"setupFiles": [
+			"<rootDir>/test/tests_utilities/setup.spec.ts"
 		],
 		"testPathIgnorePatterns": [
 			"<rootDir>/test/components/MiniCart",
