+++ modules/apps/commerce/commerce-frontend-js/src/main/resources/META-INF/resources/index.js
@@ -27,6 +27,7 @@ export {
 	MiniCart,
 	MiniCartContext,
 	Modal,
+	Multishipping,
 	Price,
 	ProductOptionCheckbox,
 	ProductOptionCheckboxMultiple,
