+++ modules/apps/commerce/commerce-frontend-js/test/components/AddDeliveryGroupButton/AddDeliveryGroupButton.spec.tsx
@@ -10,7 +10,7 @@ import fetchMock from 'fetch-mock';
 import {
 	ICountryAPIResponse,
 	IPostalAddressAPIResponse,
-} from '../../../src/main/resources/META-INF/resources/js/multishipping/Types';
+} from '../../../src/main/resources/META-INF/resources/components/multishipping/Types';
 
 import '@testing-library/jest-dom/extend-expect';
 import {
@@ -23,8 +23,8 @@ import {
 import React from 'react';
 import {act} from 'react-dom/test-utils';
 
-import AddDeliveryGroupButton from '../../../src/main/resources/META-INF/resources/js/multishipping/AddDeliveryGroupButton';
-import {setFieldValue} from '../../utils/utils.spec';
+import AddDeliveryGroupButton from '../../../src/main/resources/META-INF/resources/components/multishipping/AddDeliveryGroupButton';
+import {setFieldValue} from '../../tests_utilities/utils.spec';
 
 interface ILocators {
 	addDeliveryGroupButton: HTMLButtonElement;
