+++ modules/apps/commerce/commerce-frontend-js/test/components/DeliveryGroupHeaderCell/DeliveryGroupHeaderCell.spec.tsx
@@ -10,15 +10,15 @@ import fetchMock from 'fetch-mock';
 import {
 	ICountryAPIResponse,
 	IPostalAddressAPIResponse,
-} from '../../../src/main/resources/META-INF/resources/js/multishipping/Types';
+} from '../../../src/main/resources/META-INF/resources/components/multishipping/Types';
 
 import '@testing-library/jest-dom/extend-expect';
 import {RenderResult, cleanup, render, waitFor} from '@testing-library/react';
 import React from 'react';
 import {act} from 'react-dom/test-utils';
 
-import DeliveryGroupHeaderCell from '../../../src/main/resources/META-INF/resources/js/multishipping/DeliveryGroupHeaderCell';
-import {setFieldValue} from '../../utils/utils.spec';
+import DeliveryGroupHeaderCell from '../../../src/main/resources/META-INF/resources/components/multishipping/DeliveryGroupHeaderCell';
+import {setFieldValue} from '../../tests_utilities/utils.spec';
 
 interface ILocators {
 	actionsButton: HTMLButtonElement;
