+++ modules/apps/commerce/commerce-frontend-js/test/components/DeliveryGroupModal/DeliveryGroupModal.spec.tsx
@@ -11,15 +11,15 @@ import {
 	ICountryAPIResponse,
 	IPostalAddress,
 	IPostalAddressAPIResponse,
-} from '../../../src/main/resources/META-INF/resources/js/multishipping/Types';
+} from '../../../src/main/resources/META-INF/resources/components/multishipping/Types';
 
 import '@testing-library/jest-dom/extend-expect';
 import {RenderResult, cleanup, render, waitFor} from '@testing-library/react';
 import React from 'react';
 import {act} from 'react-dom/test-utils';
 
-import DeliveryGroupModal from '../../../src/main/resources/META-INF/resources/js/multishipping/DeliveryGroupModal';
-import {setFieldValue} from '../../utils/utils.spec';
+import DeliveryGroupModal from '../../../src/main/resources/META-INF/resources/components/multishipping/DeliveryGroupModal';
+import {setFieldValue} from '../../tests_utilities/utils.spec';
 
 interface ILocators {
 	addressCountrySelect: HTMLSelectElement;
