+++ modules/apps/commerce/commerce-order-content-web/package.json
@@ -2,37 +2,18 @@
 	"dependencies": {
 		"@clayui/alert": "3.111.0",
 		"@clayui/button": "3.116.0",
-		"@clayui/drop-down": "3.121.0",
-		"@clayui/empty-state": "3.116.0",
 		"@clayui/form": "3.119.0",
 		"@clayui/icon": "3.111.0",
 		"@clayui/label": "3.111.0",
-		"@clayui/link": "3.111.0",
-		"@clayui/loading-indicator": "3.111.0",
-		"@clayui/management-toolbar": "3.111.0",
 		"@clayui/modal": "3.119.0",
 		"@clayui/multi-step-nav": "3.121.0",
-		"@clayui/pagination-bar": "3.121.0",
 		"@clayui/sticker": "3.111.0",
-		"@clayui/table": "3.111.0",
-		"@clayui/tooltip": "3.119.0",
 		"classnames": "2.3.1",
 		"commerce-frontend-js": "*",
-		"fetch-mock": "5.13.1",
 		"frontend-js-web": "*",
-		"lodash": "4.17.21",
 		"moment": "2.29.4",
 		"prop-types": "15.7.2",
-		"react": "16.12.0",
-		"react-dom": "16.12.0"
-	},
-	"jest": {
-		"setupFiles": [
-			"<rootDir>/test/utils/setup.spec.ts"
-		],
-		"testPathIgnorePatterns": [
-			"<rootDir>/test/utils/"
-		]
+		"react": "16.12.0"
 	},
 	"main": "js/index.js",
 	"name": "@liferay/commerce-order-content-web",
