+++ modules/apps/commerce/commerce-order-content-web/src/main/resources/META-INF/resources/js/index.js
@@ -15,7 +15,6 @@ export {default as editCommerceOrder} from './editCommerceOrder';
 export {default as editCommerceReturn} from './editCommerceReturn';
 export {default as editCommerceReturnItem} from './editCommerceReturnItem';
 export {default as InlineEditableOrderField} from './inline_editable_order_field/InlineEditableOrderField';
-export {default as Multishipping} from './multishipping/Multishipping';
 export {default as newView} from './newView';
 export {default as createCommerceCart} from './order_data_set/createCommerceCart';
 export {default as OrderDataSetPropsTransformer} from './order_data_set/propsTransformer';
